updated field

+++ Lab-11/KenaKata/src/main/java/kenakata/catalog/capabilities/Weighable.java
@@ -1,5 +1,5 @@
 package kenakata.catalog.capabilities;
 
 public interface Weighable {
-    double weightInGrams();
+    int weightInGrams();
 }
